Hamza dev

+++ controllers/userControllers.js
@@ -89,28 +89,30 @@ exports.delete = async (req, res) => {
     if (user) {
       const reviewedId = [];
       // remove user.Reviews from book.Reviews and delete all user.review
-      for (let i = 0; i < user.reviews.length; i++) {
-        reviewedId.push(user.reviews[i]);
-      }
-      console.log(reviewedId);
-      if (reviewedId.length>0) {
-        for (let j = 0; j < reviewedId.length; j++) {
-          const review = await Review.findById(reviewedId[j])  ;
-         // remove user.Reviews from book.Reviews
-           await Book.findByIdAndUpdate(
-            review.book ,
-            { $pull: { reviews: reviewedId[j] } },
-            {
-              new: true,
-            }
-          );
+      
+      if (user.reviews) {
+        for (let i = 0; i < user.reviews.length; i++) {
+          reviewedId.push(user.reviews[i]);
+        }
+        if (reviewedId.length > 0) {
+          for (let j = 0; j < reviewedId.length; j++) {
+            const review = await Review.findById(reviewedId[j]);
+            // remove user.Reviews from book.Reviews
+            await Book.findByIdAndUpdate(
+              review.book,
+              { $pull: { reviews: reviewedId[j] } },
+              {
+                new: true,
+              }
+            );
   
-           await Review.findByIdAndRemove(reviewedId[j]);
-       
-  
+            await Review.findByIdAndRemove(reviewedId[j]);
+          }
         }
       }
-      fs.unlinkSync("uploads/users/" + user.image);
+      if (user.image != "http://localhost:3000/uploads/users/download.jpeg") {
+        fs.unlinkSync("uploads/users/" + user.image);
+      }
       await User.findByIdAndRemove(req.params.id);
       res.json({ message: "User been deleted successfully" });
     } else {
@@ -167,7 +169,7 @@ exports.update = async (req, res) => {
 // affect a favour book to a user using book Id conroller
 exports.affectFavBook = async (req, res) => {
   try {
-    await User.findByIdAndUpdate(
+  const user =  await User.findByIdAndUpdate(
       req.user.id,
       { $push: { favoritbooks: req.params.idbook } },
       {
@@ -176,6 +178,7 @@ exports.affectFavBook = async (req, res) => {
     );
     res.status(200).json({
       message: "book added successfully to Favorits . ",
+      user : user 
     });
   } catch (error) {
     console.log(error);
@@ -185,8 +188,8 @@ exports.affectFavBook = async (req, res) => {
 // désaffact fav book from user controller
 exports.removeFavBook = async (req, res) => {
   try {
-    await User.findByIdAndUpdate(
-      req.params.iduser,
+    const user = await User.findByIdAndUpdate(
+      req.user.id,
       { $pull: { favoritbooks: req.params.idbook } },
       {
         new: true,
@@ -194,6 +197,7 @@ exports.removeFavBook = async (req, res) => {
     );
     res.status(200).json({
       message: "Book has been removed from you <3 list .  ",
+      user : user 
     });
   } catch (error) {
     console.log(error);
